LogisticsHackathon/ml-service/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
flask==3.0.3
flask-cors==4.0.0
pydantic==2.10.6
opencv-python==4.11.0.86
pytesseract==0.3.13
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
